Update index.html
<div class="col-6 col-12-medium">
						<h3><hr>
						"Editing with Sandra was fantastical from beginning to end; fantastical being part of my book title, a word in her bio and our experience working together! She is professional, talented and inspiring. 

Sandi was outstanding at finding the voice of the story and the writer, making sure they were both consistent and clear throughout. She approached my complex and abstract storyline with an open mind and fully embraced it, honing it into something so much more than when I thought it was first “done”. The way she inspected and shifted the details ultimately turned my rough stone into a polished jewel.

Sandra's work style provided me the perfect combination of feedback, instruction, boundaries and encouragement, making her an invaluable addition to my team. This is only the beginning! I look forward to working with her on the rest of my books in the series.”
						</h3>
								
							<h3>
								—Terra Nicolle, author of <i>Spiraling Up</i>
							</h3><hr>
						
					</div>